requirements.txt
selenium
python-dotenv
fastapi
uvicorn
python-multipart
jinja2
python-crontab
sqlalchemy
bcrypt
python-jose[cryptography]
gunicorn
uvicorn[standard]
webdriver-manager
